Create et3c
Today, with the advancement of blockchain and Internet of Things (IoT) technologies, there is a growing need for authentication protocols and key exchanges to facilitate communication between these diverse technologies. Authentication and key exchange protocols are designed using symmetric and asymmetric encryption methods, each with varying computational costs. In IoT systems, where memory and computation power are often limited, researchers are actively seeking lightweight protocol designs to minimize computational burdens. Traditionally, the calculation of computational and communication costs for protocols has been a manual process, which is prone to human error. In this paper, we propose the E3C tool, which automates the calculation of computation and communication costs for authentication and key exchange protocols. E3C enables the comparison of multiple protocols in terms of their communication and processing costs, presenting the results in separate charts. By comparing the processing and communication costs of both classical and modern protocols, both manually and using E3C, we demonstrate that E3C achieves 99.99% accuracy in calculating the costs of authentication and key exchange protocols.